__app__/requirements.txt
azure-functions
azure-storage-blob
PyPDF2
pandas